Code QOL, updating objects returned from predictions
Progressively trimming unused snippets of code that are redundant or unnecesary, as well as updating the prediction_result objects to be self-contained per individual prediction instead of a master dictionary. Now predict() returns a dictionary of individual prediction_results objects, and those have their own variables with prediction information as well as methods to explore the results (the same functions but implemented as methods). Documentation should be updated... sooner than later.